Refactor non-route code into services, and move root services into folder
